+++ packages/tandem-designer/src/components/Main/DesignMode/RightSidebar/CSSInspector/index.pc
@@ -722,3 +722,16 @@
     </span>
   </div>
 </div>
+
+<!--
+  @frame { height: 269, width: 360, x: -256, y: -562 }
+-->
+<div>
+  <style>
+    background: rgb(0,0,0);
+    color: white;
+    padding: 16px;
+  </style>
+
+  but it might not be
+</div>
